resource-service/app/requirements.txt
requests
pyjwt
cryptography
lambda-cache
joblib
diskcache
